refactor registerPage into 1 component/page
